rename entity files, update imports
