+++ api/src/dto/info.dto.ts
@@ -1,4 +1,4 @@
 import { InfoDTO } from 'nestjs-base-service';
-import { User } from 'modules/users/user.entity';
+import { User } from 'modules/users/user.api.entity';
 
 export type AppInfoDTO = InfoDTO<User>;
