+++ api/src/modules/countries/countries.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get, Param, UseGuards } from '@nestjs/common';
-import { Country, CountryResult } from './country.entity';
+import { Country, CountryResult } from './country.api.entity';
 import { CountriesService } from './countries.service';
 
 import JSONAPISerializer = require('jsonapi-serializer');
